Fixed and documented mustache.sh submodule.
Closes #10.

+++ bootstrap.sh
@@ -1,19 +1,22 @@
-echo "Paste the contents of ~/production.pem here and press ^D:" >&2
-cat >"$HOME/production.pem"
-chmod 600 "$HOME/production.pem"
+if [ ! -f "$HOME/production.pem" ]
+then
+	echo "Paste the contents of ~/production.pem here and press ^D:" >&2
+	cat >"$HOME/production.pem"
+	chmod 600 "$HOME/production.pem"
+fi
 
-apt-get -q update
+sudo apt-get -q update
 
-apt-get -q -y install \
-	build-essential \
-	git-core \
-	python python-setuptools \
-	ruby ruby-dev rubygems \
+sudo apt-get -q -y install \
+	"build-essential" \
+	"git-core" \
+	"python" "python-setuptools" \
+	"ruby" "ruby-dev" "rubygems" \
 
-gem install fpm
+sudo gem install --no-rdoc --no-ri "fpm"
 
-mkdir "$HOME/work"
+mkdir -p "$HOME/work"
 cd "$HOME/work"
-git clone git://github.com/devstructure/blueprint.git
+[ -d "blueprint" ] || git clone "git://github.com/devstructure/blueprint.git"
 (cd "blueprint" && git submodule update --init)
-git clone git://github.com/rcrowley/ronn.git -b dots
+[ -d "ronn" ] || git clone "git://github.com/rcrowley/ronn.git" -b "dots"
